repo update

+++ src/api/services/task.services.js
@@ -3,13 +3,13 @@
 
 const Task = require('../models/task.model')
 const Group = require('../models/group.model')
-const groupServices = require('./group.services')
+const { startOfDay, endOfDay } = require('date-fns');
 
 
 // Create a single task in a group.
-const createTask = async(userID, {groupID, taskName, taskDesc, startDate, endDate, attachedFiles}) => 
+const createTask = async(userID, groupID, taskName, taskDesc, isHabit, startDate, endDate, attachedFiles) => 
 {
-    console.log("ID: "+groupID)
+    console.log(taskName)
     try 
     {
         const tempTask = new Task
@@ -18,11 +18,10 @@ const createTask = async(userID, {groupID, taskName, taskDesc, startDate, endDat
                 assigner: userID,
                 group: groupID, 
                 name:taskName, 
-                taskDesc, 
-                 
+                description:taskDesc, 
+                isHabit,
                 startDate, 
                 endDate, 
-                 
                 attachedFiles
             }
         )
@@ -54,12 +53,16 @@ const readAllTasks = async() =>
 
 
 // Read a task using its unique ID.
-const readTask = async(taskID) => 
+const readTask = async(taskID, userID) => 
 {
     try 
     {
+        let isDone = false
         const task = await Task.findById(taskID)
-        return task
+
+        if (task.userCompleted.includes(userID))
+            isDone = true
+        return {task, isDone}
     } 
     catch (err) 
     {
@@ -70,19 +73,19 @@ const readTask = async(taskID) =>
 
 
 // Read the tasks of a single group.
-const readGroupTasks = async(groupID) =>
+const readGroupTasks = async(groupID, userID) =>
 {
     try
     {
-        const groupTasksId = await Group.findById(groupID, {_id:0, tasks: 1})
-        const tasks = await Task.find({_id: { $in: groupTasksId.tasks }}, {name:1, status:1, attachedFiles:1, whoCompletedTask:1})
-        let temp = []
+        // Get all task for this group
+        const groupTaskIDs = await Group.findById(groupID, {tasks: 1})
+        const tasks = await Task.find({_id: { $in: groupTaskIDs.tasks }}, {name:1, isHabit:1, attachedFiles:1, userCompleted:1})
+        
         const completeTasks = []
         const incompleteTasks = []
         for (let task in tasks)
         {
-            temp = task.whoCompletedTask || []
-            if (temp.length > 0)
+            if (tasks[task].userCompleted.includes(userID))
                 completeTasks.push(tasks[task])
             else
                 incompleteTasks.push(tasks[task])
@@ -102,15 +105,53 @@ const readTodayTasks = async(userID) =>
 {
     try
     {
-        const userGroups = await Group.find({joinedUsers: userID})
-        console.log(userGroups)
-        return userGroups
+        const userGroups = await Group.find({joinedUsers: userID}, {_id:1})
+        
+        const start = startOfDay(new Date())
+        const end = endOfDay(new Date())
+        let idArr = []
+        
+
+        const todayTasks = await Task.find({group: {$in: userGroups}, endDate: {$gte: start, $lte:end}}, {_id:1, userCompleted:1, group:1})
+        
+        for (let taskObj in todayTasks)
+        {
+            idArr.push(todayTasks[taskObj]._id)
+        }
+
+        let groupPictures = []
+        for (id in idArr)
+            {
+                let groupPicture = await Group.find({tasks:idArr[id]}, {_id:0, groupPhoto:1})
+                groupPictures.push(groupPicture[0].groupPhoto)
+            }
+
+        let completeTasks = []
+        let incompleteTasks = []
+        for (let task in todayTasks)
+        {
+            
+            if (todayTasks[task].userCompleted[0] !== undefined)
+                completeTasks.push(todayTasks[task])
+            else
+                incompleteTasks.push(todayTasks[task])
+        }
+        complete = completeTasks.length
+        incomplete = incompleteTasks.length
+        
+        let doneDay = false
+        if (complete !== 0 && incomplete === 0)
+            doneDay = true
+
+        
+        return {todayTasks, groupPictures, complete, incomplete, doneDay}
     }
     catch (err)
     {
         console.log("Read today tasks error" + err)
     }
-}
+} 
+
 
 
 // Update a task in a group using the task's unique ID.
@@ -124,6 +165,7 @@ const updateTask = async(taskID, updates) =>
             {
                 name: updates.taskName,
                 description: updates.taskDesc,
+                isHabit: updates.isHabit,
                 startDate: updates.startDate,
                 endDate: updates.endDate,
                 attachedFiles: updates.attachedFiles
@@ -139,12 +181,47 @@ const updateTask = async(taskID, updates) =>
 }
 
 
+// Mark task as done
+const markTask = async(userID, taskID) =>
+{
+    try
+    {
+        let isCompleted = await Task.find({_id: taskID, userCompleted: {$in: userID}}, {name:1})
+        if (isCompleted == "")
+        {
+            const group = await Task.findById(taskID)
+            const isParticipant = await Group.find({_id: group._id, joinedUsers: {$in: userID}})
+            if (isParticipant)
+            {
+                isCompleted = await Task.findByIdAndUpdate(taskID, {$push: {userCompleted: userID}})
+                if (isCompleted)
+                    return true
+            }
+        }
+        return false
+    }
+    catch (err)
+    {
+        console.log("Mark task error: " + err)
+    }
+}
+
 
 // Delete a single task from a group using the task's unique ID.
 const deleteTask = async (taskID) => 
 {
     try 
     {
+        const task = await Task.findById(taskID, {_id:1, group:1})
+        if (!task)
+            return false
+        
+        const groupID = task.group
+        console.log("group id:" + task)
+        const group = await Group.findByIdAndUpdate(groupID, {$pullAll: {tasks: [taskID]}})
+        if (!group)
+            return false
+
         const deleteCount = await Task.findByIdAndDelete(taskID)
         return deleteCount
     } 
@@ -154,20 +231,6 @@ const deleteTask = async (taskID) =>
     }
 }
 
-
-
-// Search through tasks within a group.
-const searchTasks = async() => 
-{
-    try 
-    {
-
-    } 
-    catch (err) 
-    {
-        console.log('search task error: ' + err)
-    }
-}
 
 
 
@@ -180,6 +243,6 @@ module.exports =
     readGroupTasks,
     readTodayTasks,
     updateTask,
-    deleteTask,
-    searchTasks
+    markTask,
+    deleteTask
 }
